update strings and separate widgets to files
